+++ configure.ac
@@ -2511,6 +2511,7 @@ OUTPUT_FILES="Makefile \
     xbmc/guilib/Makefile \
     xbmc/input/linux/Makefile \
     xbmc/interfaces/Makefile \
+    xbmc/interfaces/builtins/Makefile \
     xbmc/network/Makefile \
     xbmc/network/httprequesthandler/python/Makefile \
     xbmc/network/upnp/Makefile \
