+++ xbmc/addons/GUIDialogAddonInfo.cpp
@@ -40,7 +40,7 @@
 #include "addons/AddonInstaller.h"
 #include "pvr/PVRManager.h"
 #include "Util.h"
-#include "interfaces/Builtins.h"
+#include "interfaces/builtins/Builtins.h"
 
 #define CONTROL_BTN_INSTALL          6
 #define CONTROL_BTN_ENABLE           7
