+++ xbmc/interfaces/Makefile.in
@@ -1,5 +1,4 @@
 SRCS  = AnnouncementManager.cpp
-SRCS += Builtins.cpp
 
 LIB = interfaces.a
