+++ xbmc/interfaces/builtins/Makefile.in
@@ -1,0 +1,6 @@
+SRCS += Builtins.cpp
+     
+LIB=builtins.a
+
+include ../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
