+++ xbmc/interfaces/json-rpc/PlayerOperations.cpp
@@ -27,7 +27,7 @@
 #include "input/Key.h"
 #include "GUIUserMessages.h"
 #include "pictures/GUIWindowSlideShow.h"
-#include "interfaces/Builtins.h"
+#include "interfaces/builtins/Builtins.h"
 #include "PartyModeManager.h"
 #include "ApplicationMessenger.h"
 #include "FileItem.h"
