+++ xbmc/interfaces/json-rpc/SystemOperations.cpp
@@ -20,7 +20,7 @@
 
 #include "SystemOperations.h"
 #include "ApplicationMessenger.h"
-#include "interfaces/Builtins.h"
+#include "interfaces/builtins/Builtins.h"
 #include "utils/Variant.h"
 #include "powermanagement/PowerManager.h"
